move rendering provider configuration to the provider, to allow other providers to define an own one
