+++ plugins/GeoIp2/angularjs/geoip2-updater/geoip2-updater.controller.js
@@ -5,11 +5,11 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 (function () {
-    angular.module('piwikApp').controller('LocationProviderUpdaterController', LocationProviderUpdaterController);
+    angular.module('piwikApp').controller('Geoip2UpdaterController', Geoip2UpdaterController);
 
-    LocationProviderUpdaterController.$inject = ['piwikApi', '$window'];
+    Geoip2UpdaterController.$inject = ['piwikApi', '$window'];
 
-    function LocationProviderUpdaterController(piwikApi, $window) {
+    function Geoip2UpdaterController(piwikApi, $window) {
         // remember to keep controller very simple. Create a service/factory (model) if needed
         var self = this;
 
@@ -25,7 +25,7 @@
 
             piwikApi.withTokenInUrl();
             piwikApi.post({
-                module: 'UserCountry',
+                module: 'GeoIp2',
                 action: action,
                 'continue': cont ? 1 : 0
             }, data).then(function (response) {
@@ -135,7 +135,7 @@
             piwikApi.withTokenInUrl();
             piwikApi.post({
                 period: this.updatePeriod,
-                module: 'UserCountry',
+                module: 'GeoIp2',
                 action: 'updateGeoIPLinks'
             }, {
                 loc_db: this.locationDbUrl,
