+++ plugins/UserCountry/templates/adminIndex.twig
@@ -14,20 +14,8 @@
 <div piwik-content-block content-title="{{ 'UserCountry_LocationProvider'|translate|e('html_attr') }}">
 <div piwik-location-provider-selection="{{ currentProviderId|e('html_attr') }}">
 
-    {% if not isThereWorkingProvider and hasGeoIp2Provider %}
-        <h3 style="margin-top:0;">{{ 'UserCountry_HowToSetupGeoIP'|translate }}</h3>
-        {% if dbipLiteUrl|default is not empty %}
-        <p>{{ 'UserCountry_HowToSetupGeoIPIntro'|translate }}</p>
-        <ul style="list-style:disc !important;margin-left:2em;">
-            <li style="list-style-type: disc !important;">{{ 'UserCountry_HowToSetupGeoIP_Step1'|translate('<a rel="noreferrer noopener" href="'~dbipLiteUrl~'">','</a>','<a rel="noreferrer noopener" target="_blank" href="http://db-ip.com/?refid=mtm">','</a>')|raw }}</li>
-            <li style="list-style-type: disc !important;">{{ 'UserCountry_HowToSetupGeoIP_Step2'|translate("'"~dbipLiteFilename~"'",'<strong>','</strong>','<strong>'~dbipLiteDesiredFilename~'</strong>')|raw }}</li>
-            <li style="list-style-type: disc !important;">{{ 'UserCountry_HowToSetupGeoIP_Step3'|translate('<strong>','</strong>','<span style="color:green"><strong>','</strong></span>')|raw }}</li>
-            <li style="list-style-type: disc !important;">{{ 'UserCountry_HowToSetupGeoIP_Step4'|translate }}</li>
-        </ul>
-        {% else %}
-        <p>{{ 'UserCountry_InstallGeoIp2'|translate('<a rel="noreferrer noopener" target="_blank" href="http://db-ip.com/?refid=mtm">','</a>')|raw }}</p>
-        {% endif %}
-        <p>&nbsp;</p>
+    {% if not isThereWorkingProvider %}
+        {{ setUpGuides|raw }}
     {% endif %}
 
     <div class="row">
@@ -102,56 +90,7 @@
 </div>
 </div>
 
-{% if isInternetEnabled and hasGeoIp2Provider %}
-    <div piwik-content-block
-         content-title="{% if not geoIPDatabasesInstalled %}{{ 'UserCountry_GeoIPDatabases'|translate|e('html_attr') }}{% else %}{{ 'UserCountry_SetupAutomaticUpdatesOfGeoIP'|translate|e('html_attr') }}{% endif %}"
-         id="geoip-db-mangement">
-
-        <div piwik-location-provider-updater
-             geoip-database-installed="{% if geoIPDatabasesInstalled %}1{% else %}0{% endif %}">
-
-            {% if showGeoIPUpdateSection %}
-                {% if not geoIPDatabasesInstalled %}
-                    <div ng-show="!locationUpdater.geoipDatabaseInstalled">
-                        <div ng-show="locationUpdater.showPiwikNotManagingInfo">
-                            <h3>{{ 'UserCountry_PiwikNotManagingGeoIPDBs'|translate|e('html_attr') }}</h3>
-                            <div id="manage-geoip-dbs">
-                                <div class="row" id="geoipdb-screen1">
-                                    <div class="geoipdb-column-1 col s6">
-                                        <p>{{ 'UserCountry_IWantToDownloadFreeGeoIP'|translate|raw }}</p>
-                                    </div>
-                                    <div class="geoipdb-column-2 col s6">
-                                        <p>{{ 'UserCountry_IPurchasedGeoIPDBs'|translate('<a rel="noreferrer noopener" href="http://www.maxmind.com/en/geolocation_landing?rId=piwik">','</a>','<a rel="noreferrer noopener" href="https://db-ip.com/db/?refid=mtm">','</a>')|raw }}</p>
-                                    </div>
-                                    <div class="geoipdb-column-1 col s6">
-                                        <input type="button" class="btn"
-                                               ng-click="locationUpdater.startDownloadFreeGeoIp()"
-                                               value="{{ 'General_GetStarted'|translate }}..."/>
-                                    </div>
-                                    <div class="geoipdb-column-2 col s6">
-                                        <input type="button" class="btn"
-                                               ng-click="locationUpdater.startAutomaticUpdateGeoIp()"
-                                               value="{{ 'General_GetStarted'|translate }}..." id="start-automatic-update-geoip"/>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        <div id="geoipdb-screen2-download" ng-show="locationUpdater.showFreeDownload">
-                            <div piwik-progressbar
-                                 label="{{ ('UserCountry_DownloadingDb'|translate('<a href="'~dbipLiteUrl~'">'~dbipLiteFilename~'</a>') ~ '...')|json_encode }}"
-                                 progress="locationUpdater.progressFreeDownload">
-                            </div>
-                        </div>
-                    </div>
-                {% endif %}
-
-                {% include "@UserCountry/_updaterManage.twig" %}
-            {% else %}
-                <p class="form-description">{{ 'UserCountry_CannotSetupGeoIPAutoUpdating'|translate }}</p>
-            {% endif %}
-        </div>
-    </div>
-{% endif %}
+{{ configurations|raw }}
 
 {% endblock %}
